requirements.txt

pyspark==3.5.0
kafka-python==2.0.2
pytest==7.4.0
pytest-cov==4.1.0